bug fix in sum(), d should be double

+++ ximple-dev/com/ximpleware/FuncExpr.java
@@ -706,7 +706,7 @@ public class FuncExpr extends Expr{
 	}
 	
 	private double sum(VTDNav vn){
-	    int d=0;
+	    double d=0;
 	    if (argCount() != 1 || argumentList.e.isNodeSet() == false)
 	        throw new IllegalArgumentException("sum()'s argument count is invalid");
     	vn.push2();
